+++ src/components/InstanceFilters/InstanceFilters.js
@@ -14,6 +14,7 @@ import CheckboxFacet from '../CheckboxFacet';
 import DateRangeFilter from '../DateRangeFilter';
 import {
   getSourceOptions,
+  getSharedOptions,
   getSuppressedOptions,
   processFacetOptions,
   processStatisticalCodes,
@@ -52,6 +53,7 @@ const InstanceFilters = props => {
   const intl = useIntl();
 
   const segmentAccordions = {
+    [FACETS.SHARED]: false,
     [FACETS.EFFECTIVE_LOCATION]: false,
     [FACETS.LANGUAGE]: false,
     [FACETS.RESOURCE]: false,
@@ -69,6 +71,7 @@ const InstanceFilters = props => {
   };
 
   const segmentOptions = {
+    [FACETS_OPTIONS.SHARED_OPTIONS]: [],
     [FACETS_OPTIONS.EFFECTIVE_LOCATION_OPTIONS]: [],
     [FACETS_OPTIONS.LANG_OPTIONS]: [],
     [FACETS_OPTIONS.RESOURCE_TYPE_OPTIONS]: [],
@@ -84,6 +87,7 @@ const InstanceFilters = props => {
   };
 
   const selectedFacetFilters = {
+    [FACETS.SHARED]: activeFilters[FACETS.SHARED],
     [FACETS.EFFECTIVE_LOCATION]: activeFilters[FACETS.EFFECTIVE_LOCATION],
     [FACETS.LANGUAGE]: activeFilters[FACETS.LANGUAGE],
     [FACETS.RESOURCE]: activeFilters[FACETS.RESOURCE],
@@ -107,6 +111,9 @@ const InstanceFilters = props => {
         const commonProps = [recordValues, accum, name];
 
         switch (recordName) {
+          case FACETS_CQL.SHARED:
+            accum[name] = getSharedOptions(activeFilters[FACETS.SOURCE], recordValues);
+            break;
           case FACETS_CQL.EFFECTIVE_LOCATION:
             processFacetOptions(activeFilters[FACETS.EFFECTIVE_LOCATION], locations, ...commonProps);
             break;
@@ -167,6 +174,24 @@ const InstanceFilters = props => {
 
   return (
     <AccordionSet accordionStatus={accordions} onToggle={onToggleSection}>
+      <Accordion
+        label={<FormattedMessage id={`ui-inventory.filters.${FACETS.SHARED}`} />}
+        id={FACETS.SHARED}
+        name={FACETS.SHARED}
+        separator={false}
+        header={FilterAccordionHeader}
+        displayClearButton={activeFilters[FACETS.SHARED]?.length > 0}
+        onClearFilter={() => onClear(FACETS.SHARED)}
+      >
+        <CheckboxFacet
+          data-test-filter-instance-shared
+          name={FACETS.SHARED}
+          dataOptions={facetsOptions[FACETS_OPTIONS.SHARED_OPTIONS]}
+          selectedValues={activeFilters[FACETS.SHARED]}
+          isPending={getIsPending(FACETS.SHARED)}
+          onChange={onChange}
+        />
+      </Accordion>
       <Accordion
         label={<FormattedMessage id={`ui-inventory.filters.${FACETS.EFFECTIVE_LOCATION}`} />}
         id={FACETS.EFFECTIVE_LOCATION}